transaction: move structs into to the header
